Move sample Tile preview to debug folder
